requirements.txt
Django>=3.0,<3.1
djangorestframework>=3.11,<3.12
psycopg2>=2.7.5,<2.8.0
Pillow>=7.1.0,<7.1.2
flake8>=3.7.9,<3.8.0
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.0
pytz==2019.3
whitenoise==5.0.1